Rework project structure
Why:

As was mentioned in discussion for pr #26 the only way to avoid tricky
tags is to split sync and lwt packages.

Now we have three packages: `redis`, `redis.sync` and `redis.lwt`.
* `redis` for common module types
* `redis_sync/lwt` for concrete library implementation

Also, this change will break backward compatability (in earlier versions,
`Redis_sync` was a submodule of `Redis` and `Redis_lwt` was not).